Make HP bar prettier and less misleading
Before it would at first look like each bar represented a single health
point; really it represents a percent of the whole (a 15th of 100% HP).
So it now uses block characters. Those are less easy to count - and
that's good, since we don't want the bar to be counted, anyways! It's
only a *rough* representation of how hurt the character is.