Relax prompt keyboard controls
Checks the user inputted text for matches with the beginnings of titles.
(For example, `;s` is now enough to select `Sap`.)

If the selection is ambiguous, nothing will happen; the prompt will wait
for a less ambiguous input. (For example, `:ka` won't select anything if
the prompt contains both `Kafrizzle` and `Kaswoosh`.)