In UserController#list, don't return users whose Discord members are not joined to the game guild
This avoids a crash during setup where music roles would be added to
users who aren't in the guild.

This definitely isn't a perfect solution to people leaving the Discord
server; for example, the game (probably) crashes if one of a team member
has left the game.

See also #2 (on git.imalex.xyz).