Change around music
I'm opinionated and I think this music is better.

Did you know Git thinks that I *renamed* lonely-void.mp3, the old bitsy
song, to the totally new truth-in-the-stones.mp3 file? Because
apparently I didn't rewrite *enough* of the old song's bits for Git to
think this is a new file!